requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
jinja2
passlib[bcrypt]
python-jose
